2408186_Bandana_Yadav_JavaW4/PrimeNumber.java
package workshop4;

public class PrimeNumber {

    // Make the method static
    public static boolean isPrime(int number) {
        if (number <= 1) {
            return false; // Numbers less than or equal to 1 are not prime
        }

        // Check for divisibility from 2 to the square root of the number
        for (int i = 2; i <= Math.sqrt(number); i++) {
            if (number % i == 0) {
                return false; // If divisible, not prime
            }
        }

        return true; // Number is prime
    }
}


